vexpress, sunxi: move common MMU code to core/arch/arm/mm/core_mmu_v7.c
