+++ core/arch/arm/plat-sunxi/conf.mk
@@ -15,6 +15,7 @@ libutil_with_isoc := y
 CFG_SECURE_TIME_SOURCE_CNTPCT := y
 CFG_WITH_STACK_CANARIES := y
 CFG_SUNXI_UART ?= y
+CFG_MMU_V7_TTB ?= y
 
 include mk/config.mk
